# docs/infrastructure/api_v2/current-openapi-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Assixx API Documentation",
    "version": "1.0.0",
    "description": "Multi-Tenant SaaS Platform für Industrieunternehmen",
    "contact": { "name": "SCS-Technik", "email": "support@scs-technik.de" },
    "license": { "name": "Proprietary", "url": "https://assixx.com/license" }
  },
  "servers": [
    { "url": "https://api.assixx.com/api", "description": "Production Server" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token in the format: Bearer <token>"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "description": "Error message" },
          "message": {
            "type": "string",
            "description": "Error message (alternative field)"
          },
          "code": { "type": "string", "description": "Error code" },
          "details": {
            "type": "object",
            "description": "Additional error details"
          }
        }
      },
...
      "KvpSuggestion": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "description": "Suggestion ID" },
          "title": { "type": "string", "description": "Suggestion title" },
          "description": {
            "type": "string",
            "description": "Detailed description"
          },
          "category": {
            "type": "string",
            "enum": [
              "Sicherheit",
              "Produktivität",
              "Qualität",
              "Kosten",
              "Umwelt",
              "Sonstiges"
            ],
            "description": "Suggestion category"
          },
          "status": {
            "type": "string",
            "enum": ["pending", "approved", "implemented", "rejected"],
            "description": "Current status"
          },
          "is_shared": {
            "type": "boolean",
            "description": "Whether the suggestion is shared publicly"
          },
          "created_by": {
            "type": "integer",
            "description": "User ID who created the suggestion"
          },
          "tenant_id": { "type": "integer", "description": "Tenant ID" },
          "potential_savings": {
            "type": "number",
            "nullable": true,
            "description": "Estimated savings in EUR"
          },
          "implementation_effort": {
            "type": "string",
            "nullable": true,
            "enum": ["low", "medium", "high"],
            "description": "Effort required to implement"
          },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" },
          "comment_count": {
            "type": "integer",
            "description": "Number of comments"
          },
          "attachment_count": {
            "type": "integer",
            "description": "Number of attachments"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "category",
          "status",
          "created_by",
          "tenant_id"
        ]
      },
...
    }
  },
  "security": [{ "bearerAuth": [] }],
...
  "paths": {
...
    "/kvp": {
      "get": {
        "summary": "Get all KVP suggestions",
        "description": "Retrieve all KVP suggestions visible to the user with pagination and filtering",
        "tags": ["KVP"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": ["pending", "approved", "implemented", "rejected"]
            },
            "description": "Filter by suggestion status"
          },
          {
            "in": "query",
            "name": "category",
            "schema": { "type": "string" },
            "description": "Filter by category"
          },
          {
            "in": "query",
            "name": "search",
            "schema": { "type": "string" },
            "description": "Search in title and description"
          },
          {
            "in": "query",
            "name": "page",
            "schema": { "type": "integer", "minimum": 1, "default": 1 },
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            },
            "description": "Items per page"
          },
          {
            "in": "query",
            "name": "shared",
            "schema": { "type": "boolean" },
            "description": "Filter by shared/private status"
          }
        ],
        "responses": {
          "200": {
            "description": "KVP suggestions retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/KvpSuggestion" }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "currentPage": { "type": "integer" },
                        "totalPages": { "type": "integer" },
                        "totalItems": { "type": "integer" },
                        "itemsPerPage": { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create KVP suggestion",
        "description": "Create a new KVP suggestion",
        "tags": ["KVP"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "description", "category"],
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "Suggestion title",
                    "example": "Optimierung der Pausenzeiten"
                  },
                  "description": {
                    "type": "string",
                    "description": "Detailed description",
                    "example": "Durch versetzte Pausenzeiten können wir die Produktivität steigern..."
                  },
                  "category": {
                    "type": "string",
                    "enum": [
                      "Sicherheit",
                      "Produktivität",
                      "Qualität",
                      "Kosten",
                      "Umwelt",
                      "Sonstiges"
                    ],
                    "description": "Suggestion category"
                  },
                  "potential_savings": {
                    "type": "number",
                    "description": "Estimated savings in EUR",
                    "example": 5000
                  },
                  "implementation_effort": {
                    "type": "string",
                    "enum": ["low", "medium", "high"],
                    "description": "Effort required"
                  },
                  "is_shared": {
                    "type": "boolean",
                    "default": false,
                    "description": "Share with other users"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Suggestion created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "KVP-Vorschlag erfolgreich erstellt"
                    },
                    "suggestion": {
                      "$ref": "#/components/schemas/KvpSuggestion"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/kvp/categories": {
      "get": {
        "summary": "Get KVP categories",
        "description": "Retrieve all available KVP categories",
        "tags": ["KVP"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Categories retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "type": "string" },
                  "example": [
                    "Sicherheit",
                    "Produktivität",
                    "Qualität",
                    "Kosten",
                    "Umwelt",
                    "Sonstiges"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
...
    "/kvp/{id}": {
      "get": {
        "summary": "Get KVP suggestion by ID",
        "description": "Retrieve a specific KVP suggestion with all details",
        "tags": ["KVP"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" },
            "description": "KVP suggestion ID"
          }
        ],
        "responses": {
          "200": {
            "description": "KVP suggestion retrieved successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/KvpSuggestion" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "404": {
            "description": "Suggestion not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Vorschlag nicht gefunden"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/kvp/{id}/comments": {
      "get": {
        "summary": "Get comments for KVP suggestion",
        "description": "Retrieve all comments for a specific KVP suggestion",
        "tags": ["KVP"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" },
            "description": "KVP suggestion ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Comments retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "comments": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "integer" },
                          "comment": { "type": "string" },
                          "created_by": { "type": "integer" },
                          "created_by_name": { "type": "string" },
                          "created_at": {
                            "type": "string",
                            "format": "date-time"
                          }
                        }
                      },
                      "example": [
                        {
                          "id": 1,
                          "comment": "Sehr gute Idee! Das sollten wir umsetzen.",
                          "created_by": 42,
                          "created_by_name": "Max Mustermann",
                          "created_at": "2025-06-23T10:30:00Z"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "404": { "description": "Suggestion not found" }
        }
      },
      "post": {
        "summary": "Add comment to KVP suggestion",
        "description": "Add a new comment to a specific KVP suggestion",
        "tags": ["KVP"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" },
            "description": "KVP suggestion ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["comment"],
                "properties": {
                  "comment": {
                    "type": "string",
                    "description": "Comment text",
                    "example": "Sehr gute Idee! Das sollten wir umsetzen."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Comment added successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Kommentar erfolgreich hinzugefügt"
                    },
                    "comment": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "integer" },
                        "comment": { "type": "string" },
                        "created_at": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "404": { "description": "Suggestion not found" }
        }
      }
    },
...
  }
}